Finished up PauseSession_SP logic Created new version of Validate_ProctorSession_FN that does not hit the DB 3 times, but instead uses the already loaded TestSession Created a wrapper method around the call to SetOpportunityStatus_SP that still calls their logic currently Added tests for the ValidateProctor logic

+++ student/src/main/java/tds/student/performance/domain/TestSession.java
@@ -28,7 +28,7 @@ public class TestSession {
     private Timestamp dateEnd;
     private Timestamp dateVisited;
     private String clientName;
-    private Double proctorId;
+    private Long proctorId;
     private UUID sessionBrowser;
 
     public UUID getKey() {
@@ -87,11 +87,11 @@ public class TestSession {
         this.clientName = clientName;
     }
 
-    public Double getProctorId() {
+    public Long getProctorId() {
         return proctorId;
     }
 
-    public void setProctorId(Double proctorId) {
+    public void setProctorId(Long proctorId) {
         this.proctorId = proctorId;
     }
